+++ loader/src/stale.js
@@ -1,4 +1,6 @@
 const { Duration } = require("luxon");
+const metrics = require("./metrics");
+const { createWarningLogger } = require("./utils");
 
 /**
  * @typedef {Object} AgeStatistics
@@ -12,11 +14,22 @@ const { Duration } = require("luxon");
 
 const DEFAULT_STALE_THRESHOLD = 24 * 60 * 60 * 1000;
 
+const warn = createWarningLogger("stale");
+
 class StaleChecker {
   /** @type {Map<string, AgeStatistics>} */
   bySource = new Map();
+
   #finished = false;
 
+  /**
+   * StaleChecker tracks statistics about the age/staleness of location records
+   * found by the loader.
+   * @param {Object} options
+   * @param {Date} [options.relativeTime] When to calculate age relative to.
+   * @param {number} [options.threshold] Consider records older than this many
+   *        milliseconds to be stale.
+   */
   constructor({
     relativeTime = new Date(),
     threshold = DEFAULT_STALE_THRESHOLD,
@@ -25,7 +38,12 @@ class StaleChecker {
     this.threshold = threshold;
   }
 
-  getDataForSource(source) {
+  /**
+   * Get an {@link AgeStatistics} object for the given source.
+   * @param {string} source
+   * @returns {AgeStatistics}
+   */
+  getStatisticsForSource(source) {
     if (!this.bySource.has(source)) {
       this.bySource.set(source, {
         source,
@@ -37,10 +55,15 @@ class StaleChecker {
     return this.bySource.get(source);
   }
 
+  /**
+   * Aggregate statistics for a record and return its age in milliseconds.
+   * @param {any} record
+   * @returns {number}
+   */
   checkRecord(record) {
     if (!record.availability) return null;
 
-    const stats = this.getDataForSource(record.availability.source);
+    const stats = this.getStatisticsForSource(record.availability.source);
     const age = StaleChecker.calculateDataAge(this.relativeTime, record);
     if (age == null) return null;
 
@@ -52,46 +75,84 @@ class StaleChecker {
     return age;
   }
 
+  /**
+   * Aggregate statistics for a record and return the record if it is fresh or
+   * `null` if the record is stale.
+   * @param {any} record
+   * @returns {any|null}
+   */
   filterRecord(record) {
     const age = this.checkRecord(record);
     return age > this.threshold ? null : record;
   }
 
   /**
+   * Get data age statistics. Yields an {@link AgeStatistics} object for each
+   * source.
+   * @param {Object} [options]
+   * @param {boolean} [options.includeUnkown] Yield statistics for sources where
+   *        the age is unknown.
    * @yields {Required<AgeStatistics>}
    * @returns {Iterable<Required<AgeStatistics>>}
    */
-  *listStatistics() {
+  *listStatistics({ includeUnknown = false } = {}) {
     this.finish();
 
     for (const data of this.bySource.values()) {
-      if (data.samples.length > 0) {
+      if (includeUnknown || data.samples.length > 0) {
         yield data;
       }
     }
   }
 
+  /**
+   * Print a summary of any sources with stale data. Will also send a warning
+   * to any configured error tracking system, e.g. Sentry.
+   */
   printSummary() {
-    this.finish();
-
     const formatMillis = (s) =>
       Duration.fromMillis(Math.ceil(s / 1000) * 1000)
         .rescale()
         .toHuman();
 
-    for (const stats of this.bySource.values()) {
+    for (const stats of this.listStatistics({ includeUnknown: true })) {
       if (stats.samples.length === 0) {
-        console.error(`No age information for locations in ${stats.source}.`);
+        console.warn(`No age information for locations in ${stats.source}.`);
       } else if (stats.max > this.threshold) {
-        console.error(`${stats.source} has stale data!`);
-        console.error(`  Minimum age: ${formatMillis(stats.min)}`);
-        console.error(`  Maximum age: ${formatMillis(stats.max)}`);
-        console.error(`  Average age: ${formatMillis(stats.average)}`);
-        console.error(`  Median age:  ${formatMillis(stats.median)}`);
+        warn(
+          `${stats.source} has stale data!`,
+          {
+            min: formatMillis(stats.min),
+            max: formatMillis(stats.max),
+            average: formatMillis(stats.average),
+            median: formatMillis(stats.median),
+          },
+          true
+        );
       }
     }
   }
 
+  /**
+   * Send metrics to Datadog or whatever metrics system is configured.
+   * @param {string} prefix Prefix the metric name with this. Metrics are:
+   *        `<prefix>.age_seconds.<min|max|avg|median>`
+   */
+  sendMetrics(prefix) {
+    for (const stats of this.listStatistics()) {
+      const tags = [`source:${stats.source}`];
+      const name = [prefix, "age_seconds"].join(".");
+      metrics.gauge(`${name}.min`, stats.min / 1000, tags);
+      metrics.gauge(`${name}.max`, stats.max / 1000, tags);
+      metrics.gauge(`${name}.avg`, stats.average / 1000, tags);
+      metrics.gauge(`${name}.median`, stats.median / 1000, tags);
+    }
+  }
+
+  /**
+   * Calculate summary statistics (average, median). Generally you shouldn't
+   * call this directly; use `listStatistics()` instead.
+   */
   finish() {
     if (this.#finished) return;
